CEvent stuff & other (UNTESTED)
1. Updated arrays in CPostEffects class
2. Added inline getters in CTaskManager, CPed, CPedIntelligence, CTimer and CEntity classes
3. Updated flags in CPed and CPhysical
4. Fixed wrong data types in CGamelogic::IsPointWithinLineArea and in CBulletInfo.
5. Enum changes:
- ePedState was transferred into CPed.h
- ePrimaryTasks enums in CTaskManager.h were renamed
- eRepeatSectorList enums were renamed
- eDecisionMakerType enums in CDecisionMakerTypes.h were renamed
- Added missing enums in the eEventType
- Added ePedStats in CPedType.h
6. Some members in CWeapon were renamed.
7. All members in CEventGroup, CEventHandler, CAttractorScanner, CEntityScanner and CEventScanner were renamed and using correct data types now.
8. CPed and CPedIntelligence were refactored.
9. Added new classes (96 total):
CPedSaveStructure
CWeaponSaveStructure
CInterestingEvents
CNearbyFireScanner (unfinished)
CObjectPotentialCollisionScanner (unfinished)
CObjectScanner (unfinished)
CPedAcquaintanceScanner (unfinished)
CPedPotentialCollisionScanner (unfinished)
CPedScanner
CSexyPedScanner (unfinished)
CTickCounter
CVehiclePotentialCollisionScanner (unfinished)
CVehicleScanner
CEvent
CEventEditableResponse
CEventGunShot
CEventVehicleCollision
CEventPedCollisionWithPed
CEventPedCollisionWithPlayer
CEventPlayerCollisionWithPed
CEventObjectCollision
CEventBuildingCollision
CEventDraggedOutCar
CEventDamage
CEventDeath
CEventDeadPed
CEventGroupEvent
CEventPotentialGetRunOver
CEventPotentialWalkIntoVehicle
CEventPotentialWalkIntoObject
CEventPotentialWalkIntoPed
CEventVehicleHitAndRun
CEventPedEnteredMyVehicle
CEventRevived
CEventChatPartner
CEventSexyPed
CEventSexyVehicle
CEventPedToChase
CEventPedToFlee
CEventAttractor
CEventScriptedAttractor
CEventVehicleToSteal
CEventAcquaintancePed
CEventAcquaintancePedRespect
CEventAcquaintancePedLike
CEventAcquaintancePedDislike
CEventAcquaintancePedHate
CEventAcquaintancePedHateBadlyLit
CEventSeenCop
CEventVehicleThreat
CEventKnockOffBike
CEventGunAimedAt
CEventScriptCommand
CEventInAir
CEventVehicleDied
CEventHealthLow
CEventHealthReallyLow
CEventLowAngerAtPlayer
CEventHighAngerAtPlayer
CEventLeaderEnteredCarAsDriver
CEventLeaderExitedCarAsDriver
CEventLeaderQuitEnteringCarAsDriver
CEventOnFire
CEventVehicleOnFire
CEventPotentialWalkIntoBuilding
CEventHitByWaterCannon
CEventInWater
CEventStuckInAir
CEventInteriorUseInfo
CEventSignalAtPed
CEventLeanOnVehicle
CEventPassObject
CEventNewGangMember
CEventDontJoinPlayerGroup
CEventCreatePartnerTask
CEventCopCarBeingStolen
CEventVehicleDamage
CEventVehicleDamageCollision
CEventVehicleDamageWeapon
CEventSpecial
CEventGotKnockedOverByCar
CEventCarUpsideDown
CEventPotentialWalkIntoFire
CEventFireNearby
CEventGunShotWhizzedBy
CEventSoundLoud
CEventSoundQuiet
CEventSeenPanickedPed
CEventAreaCodes
CEventLeaderEntryExit
CEventPlayerCommandToGroup
CEventPlayerCommandToGroupAttack
CEventPlayerCommandToGroupGather
CEventEscalator
CEventDanger
CEventHandlerHistory

+++ plugin_sa/game_sa/CBulletInfo.h
@@ -15,7 +15,7 @@ class PLUGIN_API CBulletInfo {
 public:
     eWeaponType m_nWeaponType;
     CEntity *m_pCreator;
-    int      m_nDestroyTime;
+    float    m_nDestroyTime;
     bool     m_bExists;
 private:
     char _pad0D[3];
